Implement a bare-bones velocity calculation based on IMU values.
This is obviously terrible. Bias error will accumulate over timing,
leading to a huge error after just some time. In the future, we should
combine this value and the car odometry via a Kalman filter, but right now
it has to do. Gazebo simulation seems to have especially high noise
values, we might get lucky on the real car.